tests/wpt/tests/fledge/tentative/fetch-ad-auction-headers-insecure-context.tentative.http.html

<!doctype html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script>
    promise_test(async t => {
      return promise_rejects_dom(t, 'NotAllowedError',
        fetch('./resources/empty.html', {adAuctionHeaders: true}));
    }, 'test fetch(<url>, {adAuctionHeaders: true}) in insecure context');
  </script>
</body>
